Changes because Godot
Those changes are automatically applied every time I'm opening the editor and running game and tests. This is exhausting to track and constant discarding, so I decided to commit as they aren't breaking anything (currently). For some reason 'world.tscn' settings are ignored, probably because I hold those in the containers. Other changes are only the white spaces, as Godot's broken code editor changes all of this automatically, even thought it's not my code...